Change DropItem's position from i32 to f64
Because we have higher time precision, we also need higher
position precision to store moving objects' (DropItems) states.

Now, it is not trivial decision because i32 did not have floating point
precision issues. We could have fixed point value such as multiplying
subpixel positions with minimal possible precision like 4 if we would like
to move in units of 1/4 pixels. But it would be limited in that we wouldn't
be able to fine-tune speed later. In the end we concluded that
having flexibility of f64 is greater than having fixed point's predictability.

Note that predictability is not the same as determinism, which means
you can reproduce the same value given the same input. Floating point
arithmetic is deterministic, but it's hardly predictable to the programmers
in the least significant bits.
If we will implement network games, determinism is important to keep
states in sync among clients, but it's not necessarily be predictable.

I wanted to consult Factorio's implementation, but it's not clear whether they
use floating point to store positions internally. The Lua interface exposes
position type in tile unit, but it's obviously downsampled from item's positions
that can be more precise than tiles.

+++ src/items.rs
@@ -247,8 +247,8 @@ pub(crate) fn render_drop_item_gl(
     state: &FactorishState,
     gl: &GL,
     item_type: &ItemType,
-    x: i32,
-    y: i32,
+    x: f64,
+    y: f64,
 ) -> Result<(), JsValue> {
     render_drop_item_mat_gl(
         state,
